Algoritms scheme added.

+++ HomeworksAndLessons/SeminarHomeworks/Seminar_006/Ex_041/Program.cs
@@ -7,11 +7,11 @@ void Main()
     double temporary = 0;
     int count = 0;
     Console.Write("Сколько чисел вы хотите ввести? ");
-    m = int.Parse(Console.ReadLine());
+    m = double.Parse(Console.ReadLine());
     for (int i = 0; i < m; i++)
     {
         Console.Write($"Введите число {i + 1} (из {m}):");
-        temporary = int.Parse(Console.ReadLine());
+        temporary = double.Parse(Console.ReadLine());
         if (temporary > 0) {count++;}
     }
     Console.Write($"Введено чисел больше 0:  ({count})");
