+++ HomeworksAndLessons/SeminarHomeworks/Seminar_006/Ex_043/Program.cs
@@ -4,7 +4,7 @@
 
 double ArgumentEnter()
 {
-    double temporary = int.Parse(Console.ReadLine());
+    double temporary = double.Parse(Console.ReadLine());
     Console.WriteLine();
     return temporary;
 }
